This PR introduces several related optimizations:
1. Introduce `strippedBucket` struct/`strippedAt` method on `floatBucketIterator` which allows to skip bound calculations (which are relatively expensive `getBound` calls) in use cases where bound values are not needed (such as `DetectReset`).
2. Reduce expensive `getBound` calls from `floatBucketIterator.Next()` (if the bound has crossed the threshold once, it will exceed it for higher buckets as well).
3. Avoid allocating `floatBucketIterator` objects on heap.
4. Switch `baseBucketIterator` methods to use a pointer receiver to avoid copying of heavy struct on each invocation (note - this does not cause any additional heap allocations).

Added two benchmarks:
1. `BenchmarkFloatHistogramAllBucketIterator ` - to benchmark the performance of `AllBucketIterator` which is used all over query engine and also for histogram response marshalling.
2. `BenchmarkFloatHistogramDetectReset ` - to benchmark the method which benefits from `strippedBucket` optimization (in addition to other changes).

The result is **46.59% to 84.52%** performance improvement; complete elimination of heap allocations from `DetectReset`; 3x reduction of number of heap allocations from `AllBucketIterator`:

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/model/histogram
                                   │    before    │                after                │
                                   │    sec/op    │   sec/op     vs base                │
FloatHistogramAllBucketIterator-10    76.83µ ± 0%   41.04µ ± 0%  -46.59% (p=0.000 n=20)
FloatHistogramDetectReset-10         141.66µ ± 0%   21.92µ ± 0%  -84.52% (p=0.000 n=20)
geomean                               104.3µ        29.99µ       -71.25%

                                   │   before   │                  after                  │
                                   │    B/op    │    B/op     vs base                     │
FloatHistogramAllBucketIterator-10   336.0 ± 0%   320.0 ± 0%    -4.76% (p=0.000 n=20)
FloatHistogramDetectReset-10         512.0 ± 0%     0.0 ± 0%  -100.00% (p=0.000 n=20)
geomean                              414.8                    ?                       ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean

                                   │   before   │                  after                  │
                                   │ allocs/op  │ allocs/op   vs base                     │
FloatHistogramAllBucketIterator-10   3.000 ± 0%   1.000 ± 0%   -66.67% (p=0.000 n=20)
FloatHistogramDetectReset-10         4.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=20)
geomean                              3.464                    ?                       ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean
```